refactor : auth.js 폴더구조 수정
auth.js 폴더 구조 수정
